finish cleaning up error management so that we don't have duplicated logic
