+++ server/filesystem/filesystem.go
@@ -61,18 +61,18 @@ func (fs *Filesystem) Path() string {
 func (fs *Filesystem) File(p string) (*os.File, Stat, error) {
 	cleaned, err := fs.SafePath(p)
 	if err != nil {
-		return nil, Stat{}, err
+		return nil, Stat{}, errors.WithStackIf(err)
 	}
 	st, err := fs.Stat(cleaned)
 	if err != nil {
-		return nil, Stat{}, err
+		return nil, Stat{}, errors.WithStackIf(err)
 	}
 	if st.IsDir() {
 		return nil, Stat{}, newFilesystemError(ErrCodeIsDirectory, nil)
 	}
 	f, err := os.Open(cleaned)
 	if err != nil {
-		return nil, Stat{}, err
+		return nil, Stat{}, errors.WithStackIf(err)
 	}
 	return f, st, nil
 }
